+++ apps/web/src/profiles/migrateLegacyAuth.ts
@@ -1,7 +1,7 @@
 import type { AuthSession } from "../auth/types"
 import { Logger } from "../debug/logger.shared"
 import { getProfileRow, putProfileRow } from "./dbHelpers"
-import { getProfile, setProfile } from "./currentProfile"
+import { getProfile, setProfile } from "./active/activeProfile"
 
 /**
  * Key under which the pre-migration reactjrx auth signal was persisted. The
